modify run_marqo to generate application package for vespa

+++ scripts/vespa_local/vespa_local.py
@@ -21,7 +21,10 @@ import textwrap
 import time
 import sys
 import yaml
-import docker
+try:
+    import docker
+except ImportError:
+    print("docker package not found. If needed, please install it using `pip install docker`.")
 import xml.etree.ElementTree as ET
 from xml.dom import minidom
 import math
@@ -104,24 +107,28 @@ class VespaLocal:
                     content_for_hosts_xml = self.get_hosts_xml_content()
                     f.write(content_for_hosts_xml)
 
-    def generate_application_package(self) -> str:
+    def generate_application_package(self, zip=True) -> str:
         # Build application package directory
         self.generate_application_package_files()
 
         # Zip up files
-        os.chdir(self.base_dir)
-        shutil.make_archive('../' + self.base_dir, 'zip', ".")
-        os.chdir("..")
-        zip_file_path = f"{self.base_dir}.zip"
+        if zip:
+            os.chdir(self.base_dir)
+            shutil.make_archive('../' + self.base_dir, 'zip', ".")
+            os.chdir("..")
+            zip_file_path = f"{self.base_dir}.zip"
 
-        if os.path.isfile(zip_file_path):
-            print(f"Zip file created successfully: {zip_file_path}")
-            # Remove the base directory
-            shutil.rmtree(self.base_dir)
-            print(f"Directory {self.base_dir} removed.")
-            return zip_file_path
-        else:
-            print("Failed to create the zip file.")
+            if os.path.isfile(zip_file_path):
+                print(f"Zip file created successfully: {zip_file_path}")
+                # Remove the base directory
+                shutil.rmtree(self.base_dir)
+                print(f"Directory {self.base_dir} removed.")
+                return zip_file_path
+            else:
+                print("Failed to create the zip file.")
+                sys.exit(1)
+        return self.base_dir
+
 
 class VespaLocalSingleNode(VespaLocal):
 
@@ -592,6 +599,15 @@ def deploy_config(args):
     os.system(f'vespa deploy "{here}"')
 
 
+def generate_application_package(args):
+    # Create instance of VespaLocal
+    # vespa_local_instance is used for starting vespa & generating application package.
+    vespa_local_instance = VespaLocalSingleNode()
+    # Generate the application package
+    file_path = vespa_local_instance.generate_application_package(zip=args.zip)
+    print(f"Application package generated at {file_path}")
+
+
 def deploy_application_package(zip_file_path: str, max_retries: int = 5, backoff_factor: float = 0.5) -> None:
     # URL and headers
     url = f"{VESPA_CONFIG_URL}/application/v2/tenant/default/prepareandactivate"
@@ -681,6 +697,12 @@ def main():
 
     clean_parser = subparsers.add_parser("stop", help="Stop local Vespa")
     clean_parser.set_defaults(func=stop)
+
+    generate_application_package_parser = subparsers.add_parser("generate-application-package", help="Generate application package")
+    generate_application_package_parser.set_defaults(func=generate_application_package)
+    generate_application_package_parser.add_argument(
+        "--zip", help="Zip the application package", action="store_true"
+    )
 
     # Parse the command-line arguments and execute the corresponding function
     args = parser.parse_args()
